+++ cpp/cluster/ClusterProperties.cc
@@ -4,6 +4,7 @@
 #include <vector>
 
 #include "ClusterProperties.h"
+#include "NeighborComputeFunctional.h"
 
 /*! \file ClusterProperties.cc
     \brief Routines for computing properties of point clusters.
@@ -11,21 +12,20 @@
 
 namespace freud { namespace cluster {
 
-/*! \param box Box containing the points
-    \param points Positions of the points making up the clusters
+/*! \param nq NeighborQuery containing the points making up the clusters
     \param cluster_idx Index of which cluster each point belongs to
-    \param Np Number of points (length of \a points and \a cluster_idx)
 
-    computeClusterProperties loops over all points in the given array and
-    determines the center of mass of the cluster as well as the gyration
-    tensor. These can be accessed after the call to compute with
-    getClusterCenters() and getClusterGyrations().
+    compute loops over all points in the given array and determines the center
+    of mass of the cluster as well as the gyration tensor. These can be
+    accessed after the call to compute with getClusterCenters() and
+    getClusterGyrations().
 */
-void ClusterProperties::compute(const box::Box& box, const vec3<float>* points,
-                                const unsigned int* cluster_idx, unsigned int Np)
+
+void ClusterProperties::compute(const freud::locality::NeighborQuery* nq,
+                                const unsigned int* cluster_idx)
 {
     // determine the number of clusters
-    const unsigned int* max_cluster_id = std::max_element(cluster_idx, cluster_idx + Np);
+    const unsigned int* max_cluster_id = std::max_element(cluster_idx, cluster_idx + nq->getNPoints());
     const unsigned int num_clusters = *max_cluster_id + 1;
 
     // allocate memory for the cluster properties and temporary arrays
@@ -37,7 +37,7 @@ void ClusterProperties::compute(const box::Box& box, const vec3<float>* points,
     // ref_pos is the first point found in a cluster, it is used as a reference
     // to compute the center in relation to, for handling of the periodic
     // boundary conditions
-    std::vector<vec3<float>> ref_pos(num_clusters, vec3<float>(0.0f, 0.0f, 0.0f));
+    std::vector<vec3<float>> ref_pos(num_clusters);
     // determine if we have seen this cluster before or not (used to initialize ref_pos)
     std::vector<bool> cluster_seen(num_clusters, false);
 
@@ -45,15 +45,14 @@ void ClusterProperties::compute(const box::Box& box, const vec3<float>* points,
     // given an array of points, the easiest way to do this is to loop over
     // all points and add the appropriate information to m_cluster_centers as
     // we go.
-    for (unsigned int i = 0; i < Np; i++)
+    for (unsigned int i = 0; i < nq->getNPoints(); i++)
     {
-        unsigned int c = cluster_idx[i];
-        vec3<float> pos = points[i];
+        const unsigned int c = cluster_idx[i];
 
         // The first time we see the cluster, mark a reference position
         if (!cluster_seen[c])
         {
-            ref_pos[c] = pos;
+            ref_pos[c] = (*nq)[i];
             cluster_seen[c] = true;
         }
 
@@ -61,8 +60,7 @@ void ClusterProperties::compute(const box::Box& box, const vec3<float>* points,
         // reference vectors as wrapped vectors relative to ref_pos. When we
         // are done, we can add the computed center to ref_pos to get the
         // center in the space frame.
-        vec3<float> delta = pos - ref_pos[c];
-        delta = box.wrap(delta);
+        const vec3<float> delta(bondVector(locality::NeighborBond(c, i), nq, ref_pos.data()));
 
         // Add the vector into the center tally so far
         m_cluster_centers[c] += delta;
@@ -76,16 +74,16 @@ void ClusterProperties::compute(const box::Box& box, const vec3<float>* points,
     {
         float s = float(m_cluster_sizes[c]);
         vec3<float> v = m_cluster_centers[c] / s + ref_pos[c];
-        m_cluster_centers[c] = box.wrap(v);
+        m_cluster_centers[c] = nq->getBox().wrap(v);
     }
 
     // Now that we have determined the centers of mass for each cluster, tally
     // up the gyration tensor. This has to be done in a loop over the points.
-    for (unsigned int i = 0; i < Np; i++)
+    for (unsigned int i = 0; i < nq->getNPoints(); i++)
     {
         unsigned int c = cluster_idx[i];
-        vec3<float> pos = points[i];
-        vec3<float> delta = box.wrap(pos - m_cluster_centers[c]);
+        vec3<float> pos = (*nq)[i];
+        vec3<float> delta = nq->getBox().wrap(pos - m_cluster_centers[c]);
 
         // get the start pointer for our 3x3 matrix
         m_cluster_gyrations(c, 0, 0) += delta.x * delta.x;