+++ cpp/locality/Voronoi.cc
@@ -15,8 +15,11 @@
 namespace freud { namespace locality {
 
 // Voronoi calculations should be kept in double precision.
-void Voronoi::compute(const box::Box &box, const vec3<double>* points, unsigned int n_points)
+void Voronoi::compute(const freud::locality::NeighborQuery* nq)
     {
+        auto box = nq->getBox();
+        auto n_points = nq->getNPoints();
+
         m_polytopes.resize(n_points);
         m_volumes.prepare(n_points);
 
@@ -43,7 +46,8 @@ void Voronoi::compute(const box::Box &box, const vec3<double>* points, unsigned 
         );
 
         for (size_t query_point_id = 0; query_point_id < n_points; query_point_id++) {
-            container.put(query_point_id, points[query_point_id].x, points[query_point_id].y, points[query_point_id].z);
+            auto point = (*nq)[query_point_id];
+            container.put(query_point_id, double(point.x), double(point.y), double(point.z));
         }
 
         voro::voronoicell_neighbor cell;